feat: refactor payload functions and add bulk payload generation
